Refactor each join type into its own class
