+++ views/JoinInterfaceView/template.html
@@ -5,19 +5,19 @@
 <svg id="overlay"></svg>
 <div id="presetSettings">
   <ol class="selectMenu">
-    <li id="CONCATENATION" class="selected">
+    <li id="Concatenation" class="selected">
       <img src="../../img/concatenate.svg"/>
       <span>No preset connections</span>
     </li>
-    <li id="THETA_JOIN">
+    <li id="ThetaJoin">
       <img src="../../img/thetaJoin.svg"/>
       <span>Preset: Connect when...</span>
     </li>
-    <li id="ORDERED_JOIN">
+    <li id="OrderedJoin">
       <img src="../../img/ordered.svg"/>
       <span>Preset: Connect in order</span>
     </li>
-    <li id="CROSS_PRODUCT">
+    <li id="CrossProduct">
       <img src="../../img/crossProduct.svg"/>
       <span>Preset: All possible connections</span>
     </li>
